requirements.txt
# AI-Powered Thesis Assistant - Production Dependencies
# Version: 2.0 - Production Grade
# Last Updated: 2024

# Core ML/NLP Libraries (Fixed versions for stability)
spacy==3.7.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
scikit-learn==1.3.2
numpy==1.24.4
scipy==1.10.1

# Advanced NLP for semantic reasoning
transformers==4.35.2
torch==2.1.1
datasets==2.14.6

# Knowledge Graph and Reasoning
neo4j==5.14.1
networkx==3.2.1
rdflib==7.0.0

# Document Parsing and OCR
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow==10.1.0
python-docx==1.1.0
python-pptx==0.6.23

# Citation and Bibliography Management
bibtexparser==1.4.0
citeproc-py==0.6.0
pybtex==0.24.0

# API Interaction
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1

# Data handling and analysis
pandas==2.1.4
python-dateutil==2.8.2
jsonschema==4.20.0

# Temporal reasoning and validation
dateparser==1.2.0
arrow==1.3.0

# Advanced text processing
nltk==3.8.1
textstat==0.7.3
langdetect==1.0.9

# Configuration and logging
pydantic==2.5.2
loguru==0.7.2
python-dotenv==1.0.0

# GUI Framework (Tkinter is built-in, but we need these for enhanced widgets)
ttkthemes==3.2.2
tkinter-tooltip==2.1.0

# Testing and quality assurance
pytest==7.4.3
pytest-asyncio==0.21.1
coverage==7.3.2
pytest-cov==4.1.0

# Development and debugging tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Performance monitoring
psutil==5.9.6
memory-profiler==0.61.0

# Additional utilities
tqdm==4.66.1
colorama==0.4.6
rich==13.7.0

